backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
python-jose
passlib[bcrypt]
annotated-doc==0.0.4

annotated-types==0.7.0

anyio==4.13.0

bcrypt==4.0.1

click==8.3.3

colorama==0.4.6

dnspython==2.8.0

ecdsa==0.19.2

email-validator==2.3.0

fastapi==0.136.1

greenlet==3.5.0

h11==0.16.0

idna==3.13

passlib==1.7.4

psycopg2-binary==2.9.12

pyasn1==0.6.3

pydantic==2.13.4

pydantic_core==2.46.4

python-dotenv==1.2.2

python-jose==3.5.0

rsa==4.9.1

six==1.17.0

SQLAlchemy==2.0.49

starlette==1.0.0

typing-inspection==0.4.2

typing_extensions==4.15.0

uvicorn==0.46.0

